Added Tokenizer and Parser for Time Entries
token_times_from_string takes in the string and returns a list of tokens, integers representing times, and a boolean value denoting if it's chargable or not.

parse_time takes in the tokens and sums up the total amount of time listed. The extra argument match_type can be left as None to count all times, or be set to True or False to count deductable or non-deductable times only.