Addressed issue #3
performed a ceiling function on the minute element in add_hours_to_time, this seems to have addressed the end time rounding error.

removed the rounding function performed in calculate_time, this addressed the actual time rounding error, though test_odd_hours needed to be adjusted to account for the unrounded output.